Add terminalizer, fix unneeded symlinks

+++ setup.sh
@@ -3,20 +3,25 @@
 # Symlinks
 q=()
 q+=('("Bash profile" .bash_profile ~/.bash_profile)')
+q+=('("Bash profile" .bashrc ~/.bashrc)')
 q+=('("Git config" .gitconfig ~/.gitconfig)')
 q+=('("Hyper config" .hyper.js ~/.hyper.js)')
 q+=('("Hosts file" hosts /etc/hosts)')
 q+=('("Acceptable shells" shells /etc/shells)')
 q+=('("Hyper config" .npmrc ~/.npmrc)')
 q+=('("Sublime Settings" sublime-settings "~/Library/Application\ Support/Sublime\ Text\ 3/Packages/User/Preferences.sublime-settings")')
+q+=('("Terminalizer config" .terminalizer.yml ~/.terminalizer/config.yml)')
+q+=('("missing file" missing ~/.missing)')
 
 for set in "${q[@]}";
 do
 	eval "parts=${set}"
 
 	if [[ ! -s "$HOME/dotfiles/${parts[1]}" ]]; then
-		ln -s "${parts[2]}" "${parts[1]}"
-		echo "created symlink for ${parts[0]}"
+		if [[ -f ${parts[2]} ]]; then
+			ln -s "${parts[2]}" "${parts[1]}"
+			echo "created symlink for ${parts[0]}"
+		fi
 	fi
 done
